refactor(sphere-client-credentials): rename file from utils.js
